+++ packages/terra-navigation-side-menu/tests/wdio/navigation-side-menu-spec.js
@@ -1,7 +1,7 @@
 Terra.describeViewports('Navigation Side Menu', ['medium'], () => {
   describe('Default navigation side menu display', () => {
     it('Default navigation side menu display', () => {
-      browser.url('/#/raw/tests/terra-navigation-side-menu/navigation-side-menu/navigation-side-menu-default');
+      browser.url('/#/raw/tests/cerner-terra-framework-docs/navigation-side-menu/navigation-side-menu-default');
       $('#test-menu').waitForDisplayed();
       Terra.validates.element('navigation side menu display', { selector: '#test-menu' });
     });
@@ -9,7 +9,7 @@ Terra.describeViewports('Navigation Side Menu', ['medium'], () => {
 
   describe('Default navigation side menu display with toolbar', () => {
     it('Default navigation side menu display with toolbar', () => {
-      browser.url('/#/raw/tests/terra-navigation-side-menu/navigation-side-menu/navigation-side-menu-default-with-toolbar');
+      browser.url('/#/raw/tests/cerner-terra-framework-docs/navigation-side-menu/navigation-side-menu-default-with-toolbar');
       $('#test-menu').waitForDisplayed();
       Terra.validates.element('navigation side menu display with toolbar', { selector: '#test-menu' });
     });
@@ -17,7 +17,7 @@ Terra.describeViewports('Navigation Side Menu', ['medium'], () => {
 
   describe('Hover navigation side menu display', () => {
     it('Hover navigation side menu display', () => {
-      browser.url('/#/raw/tests/terra-navigation-side-menu/navigation-side-menu/navigation-side-menu-default');
+      browser.url('/#/raw/tests/cerner-terra-framework-docs/navigation-side-menu/navigation-side-menu-default');
       $('#test-menu').waitForDisplayed();
       $('#test-menu #test-item-1').moveTo();
       Terra.validates.element('hover navigation side menu display', { selector: '#test-menu' });
@@ -26,7 +26,7 @@ Terra.describeViewports('Navigation Side Menu', ['medium'], () => {
 
   describe('Selected navigation side menu display', () => {
     it('Selected navigation side menu display', () => {
-      browser.url('/#/raw/tests/terra-navigation-side-menu/navigation-side-menu/navigation-side-menu-default');
+      browser.url('/#/raw/tests/cerner-terra-framework-docs/navigation-side-menu/navigation-side-menu-default');
       $('#test-menu').waitForDisplayed();
       $('#test-menu #test-item-1').moveTo();
       $('#test-menu #test-item-1').click();
@@ -37,7 +37,7 @@ Terra.describeViewports('Navigation Side Menu', ['medium'], () => {
 
   describe('isRootMenu navigation side menu display', () => {
     it('isRootMenu navigation side menu display', () => {
-      browser.url('/#/raw/tests/terra-navigation-side-menu/navigation-side-menu/navigation-side-menu-root');
+      browser.url('/#/raw/tests/cerner-terra-framework-docs/navigation-side-menu/navigation-side-menu-root');
       $('#test-menu').waitForDisplayed();
       Terra.validates.element('isRootMenu navigation side menu display', { selector: '#test-menu' });
     });