+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,14 +3,18 @@
 ## Unreleased
 
 * Changed
+  * Migrated docs from terra date input.
   * Migrated terra-collapsible-menu-view-docs
   * Migrated docs from date-picker.
   * Migrated docs from date-time-picker
 
 * Added
+  * Migrate docs from terra-navigation-side-menu.
+  * Migrated docs from terra-navigation-prompt.
   * Migrate docs from terra-navigation-layout
   * Migrate docs from terra-menu
   * Migrate docs from terra-popup
+  * Migrate docs from terra-notification-dialog.
 
 ## 1.7.0 - (March 8, 2022)