+++ packages/terra-navigation-prompt/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Added
+  * Migrated docs to terra-framework-docs.
+
+
 ## 1.62.0 - (November 16, 2021)
 
 * Changed